changes from review notes
